Move GUI files into a separate folder
