--- data/tasks.txt
@@ -1,1 +1,0 @@
-T | 0 | test
