+++ src/main/resources/view/MainWindow.fxml
@@ -15,7 +15,7 @@
             stylesheets="@../css/main.css"
             xmlns="http://javafx.com/javafx/22"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="delta.MainWindow">
+            fx:controller="delta.gui.MainWindow">
     <children>
         <TextField fx:id="userInput"
                    layoutY="558.0"
